requirements.txt
appdirs==1.4.3
asgiref==3.3.0
astroid==2.3.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.12
django-allauth==0.41.0
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-admin-interface
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
ipaddr==2.2.0
isort==4.3.21
lockfile==0.12.2
lazy-object-proxy==1.4.3
numpy==1.20.3
mccabe==0.6.1
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
Pillow==8.3.2
pyswip
pylint==2.4.4
PySocks==1.7.1
python3-openid==3.1.0
pytz==2020.4
pep517==0.8.2
progress==1.5
pyparsing==2.4.6
pytoml==0.1.21
pyit2fls==0.7.9
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
stripe==2.43.0
webencodings==0.5.1
numpy
PyQt5
Shapely
descartes
matplotlib
pyinstaller
auto-py-to-exe